Add robot mobile to scene
The robot mobile is a construction which allows the agent to play with
several different shapes (cuboid, sphere, cylinder) hanging from two
revolute joints each. The shapes can be pushed by the agent and swing
back and forth when interacted with, like a childs play mobile.
This construction gives the robota a way to touch objects which are
dynamic. This is in contrast to itself and the table, which are static
objects. I hope to observe that the preference for touching dynamic
objects will be acquired during the learning phase.