Add handles to all force contacts to robot, new collection
Now all force contacts (which are related to touch sensors) have handles
within the Environment class. This will be used to create a binary map
of touch events.

Because:
+ Previous approach was to have force sensors directly emit recorded
forces into the state
+ This approach is problematic since the scale of the proprioceptive
input and the touch input can cary greatly
+ In goal based learning this introduces more problems since achieving a
state which is comprised of forces is also way harder to reach than a
simple / binary tactile map